+++ src/main/java/gift/domain/model/dto/UserResponseDto.java
@@ -1,4 +1,6 @@
-package gift.domain.model;
+package gift.domain.model.dto;
+
+import gift.domain.model.entity.User;
 
 public class UserResponseDto {
