+++ src/main/java/gift/domain/model/dto/WishResponseDto.java
@@ -1,4 +1,4 @@
-package gift.domain.model;
+package gift.domain.model.dto;
 
 public class WishResponseDto {
