+++ src/main/java/gift/domain/model/entity/Product.java
@@ -1,4 +1,4 @@
-package gift.domain.model;
+package gift.domain.model.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
