+++ src/main/java/gift/domain/model/entity/Wish.java
@@ -1,4 +1,4 @@
-package gift.domain.model;
+package gift.domain.model.entity;
 
 
 import jakarta.persistence.Column;
