+++ src/main/java/gift/domain/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package gift.domain.repository;
 
-import gift.domain.model.User;
+import gift.domain.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
